Refactor and reorganize components and services

+++ src/components/application/reusable/Scripts/getDateFormatter.ts
@@ -1,5 +1,5 @@
 export const dateFormatter = new Intl.DateTimeFormat('pt-BR', {
-    dateStyle: 'long',
+    dateStyle: 'short',
     timeStyle: 'short',
     timeZone: 'America/Sao_Paulo'
 })
